Remove the C style roundf function. Use c++11.

+++ src/AutoKeysounds.cpp
@@ -333,8 +333,8 @@ void AutoKeysounds::Update( float fDelta )
 		iRowNow = std::max( 0, iRowNow );
 		static int iRowLastCrossed = 0;
 
-		float fBeatLast = roundf(NoteRowToBeat(iRowLastCrossed));
-		float fBeatNow = roundf(NoteRowToBeat(iRowNow));
+		float fBeatLast = std::round(NoteRowToBeat(iRowLastCrossed));
+		float fBeatNow = std::round(NoteRowToBeat(iRowNow));
 
 		bCrossedABeat = fBeatLast != fBeatNow;
